Update: styles
