+++ src/components/GoogleMaps.astro
@@ -5,7 +5,7 @@ const { data } = Astro.props;
 
 <section id={data.location} class="w-full">
   <div
-    class="overflow-hidden bg-gradient-to-br from-gray-100 to-primary/60 py-4 sm:py-16 font-sans"
+    class="overflow-hidden bg-gradient-to-br from-gray-100 via-accent/20 to-primary/40 py-4 sm:py-16 font-sans"
   >
     <div class="mx-auto max-w-7xl px-6 py-12 lg:py-0 lg:px-10">
       <div
